make untrainable arrays work on GPU (#48)
Untrainable parameters are now kept track of through a dict in each PDF.

The dict is currently informed the same way as before by passing a `NoGradArray` to the constructor. The constructor stores only NoGradArray.data (so the Array/CuArray) to make sure everything works as expected and remembers which arrays should remain constant in the `_nograd` dict. Calling `params` on the pdf will only return the trainable parameters.